new class for Landing button
